+++ bin/miden-cli/src/commands/import.rs
@@ -13,7 +13,9 @@ use miden_client::{
 };
 use tracing::info;
 
-use crate::{commands::account::maybe_set_default_account, utils::load_config_file, Parser};
+use crate::{
+    commands::account::maybe_set_default_account, errors::CliError, utils::load_config_file, Parser,
+};
 
 #[derive(Debug, Parser, Clone)]
 #[clap(about = "Import client objects. It is capable of importing notes and accounts.")]
@@ -27,27 +29,25 @@ pub struct ImportCmd {
 }
 
 impl ImportCmd {
-    pub async fn execute(&self, mut client: Client<impl FeltRng>) -> Result<(), String> {
+    pub async fn execute(&self, mut client: Client<impl FeltRng>) -> Result<(), CliError> {
         validate_paths(&self.filenames)?;
         let (mut current_config, _) = load_config_file()?;
         for filename in &self.filenames {
             let note_file = read_note_file(filename.clone());
 
             if let Ok(note_file) = note_file {
-                let note_id = client.import_note(note_file).await.map_err(|err| err.to_string())?;
+                let note_id = client.import_note(note_file).await?;
                 println!("Succesfully imported note {}", note_id.inner());
             } else {
                 info!(
                     "Attempting to import account data from {}...",
-                    fs::canonicalize(filename).map_err(|err| err.to_string())?.as_path().display()
+                    fs::canonicalize(filename)?.as_path().display()
                 );
-                let account_data_file_contents =
-                    fs::read(filename).map_err(|err| err.to_string())?;
+                let account_data_file_contents = fs::read(filename)?;
 
                 let account_id =
                     import_account(&mut client, &account_data_file_contents, self.overwrite)
-                        .await
-                        .map_err(|err| err.to_string())?;
+                        .await?;
 
                 println!("Successfully imported account {}", account_id);
 
@@ -67,7 +67,7 @@ async fn import_account(
     client: &mut Client<impl FeltRng>,
     account_data_file_contents: &[u8],
     overwrite: bool,
-) -> Result<AccountId, ClientError> {
+) -> Result<AccountId, CliError> {
     let account_data = AccountData::read_from_bytes(account_data_file_contents)
         .map_err(ClientError::DataDeserializationError)?;
     let account_id = account_data.account.id();
@@ -87,13 +87,12 @@ async fn import_account(
 // IMPORT NOTE
 // ================================================================================================
 
-fn read_note_file(filename: PathBuf) -> Result<NoteFile, String> {
+fn read_note_file(filename: PathBuf) -> Result<NoteFile, CliError> {
     let mut contents = vec![];
-    let mut _file = File::open(filename)
-        .and_then(|mut f| f.read_to_end(&mut contents))
-        .map_err(|err| err.to_string());
+    let mut _file = File::open(filename).and_then(|mut f| f.read_to_end(&mut contents))?;
 
-    NoteFile::read_from_bytes(&contents).map_err(|err| err.to_string())
+    NoteFile::read_from_bytes(&contents)
+        .map_err(|err| CliError::Client(ClientError::DataDeserializationError(err)))
 }
 
 // HELPERS
@@ -101,11 +100,11 @@ fn read_note_file(filename: PathBuf) -> Result<NoteFile, String> {
 
 /// Checks that all files exist, otherwise returns an error. It also ensures that all files have a
 /// specific extension.
-fn validate_paths(paths: &[PathBuf]) -> Result<(), String> {
+fn validate_paths(paths: &[PathBuf]) -> Result<(), CliError> {
     let invalid_path = paths.iter().find(|path| !path.exists());
 
     if let Some(path) = invalid_path {
-        Err(format!("The path `{}` does not exist", path.to_string_lossy()).to_string())
+        Err(CliError::Input(format!("The path `{}` does not exist", path.to_string_lossy())))
     } else {
         Ok(())
     }