+++ bin/miden-cli/src/commands/tags.rs
@@ -5,7 +5,7 @@ use miden_client::{
 };
 use tracing::info;
 
-use crate::Parser;
+use crate::{errors::CliError, Parser};
 
 #[derive(Default, Debug, Parser, Clone)]
 #[clap(about = "View and manage tags. Defaults to `list` command.")]
@@ -24,7 +24,7 @@ pub struct TagsCmd {
 }
 
 impl TagsCmd {
-    pub async fn execute(&self, client: Client<impl FeltRng>) -> Result<(), String> {
+    pub async fn execute(&self, client: Client<impl FeltRng>) -> Result<(), CliError> {
         match self {
             TagsCmd { add: Some(tag), .. } => {
                 add_tag(client, *tag).await?;
@@ -42,13 +42,13 @@ impl TagsCmd {
 
 // HELPERS
 // ================================================================================================
-async fn list_tags(client: Client<impl FeltRng>) -> Result<(), String> {
+async fn list_tags(client: Client<impl FeltRng>) -> Result<(), CliError> {
     let tags = client.get_note_tags().await?;
     println!("Tags: {:?}", tags);
     Ok(())
 }
 
-async fn add_tag(mut client: Client<impl FeltRng>, tag: u32) -> Result<(), String> {
+async fn add_tag(mut client: Client<impl FeltRng>, tag: u32) -> Result<(), CliError> {
     let tag: NoteTag = tag.into();
     let execution_mode = match tag.execution_mode() {
         NoteExecutionMode::Local => "Local",
@@ -64,7 +64,7 @@ async fn add_tag(mut client: Client<impl FeltRng>, tag: u32) -> Result<(), Strin
     Ok(())
 }
 
-async fn remove_tag(mut client: Client<impl FeltRng>, tag: u32) -> Result<(), String> {
+async fn remove_tag(mut client: Client<impl FeltRng>, tag: u32) -> Result<(), CliError> {
     client.remove_note_tag(tag.into()).await?;
     println!("Tag {} removed", tag);
     Ok(())