+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * [BREAKING] Made `idxdb` only usable on WASM targets (#685).
 * Added fixed seed option for web client generation (#688)
 * [BREAKING] Updated `init` command in the CLI to receive a `--network` flag (#690).
+* Improved CLI error messages (#682).
 
 ### Fixes